# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Compliance OS — GRC tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /controls:
    get:
      operationId: listControls
      tags: [controls]
      summary: List controls
      parameters:
        - name: frameworkId
          in: query
          schema:
            type: ["string", "null"]
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["number", "null"]
        - name: offset
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Controls list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Control"
  /controls/{id}:
    get:
      operationId: getControl
      tags: [controls]
      summary: Get a control
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
        "404":
          description: Not found
    patch:
      operationId: updateControl
      tags: [controls]
      summary: Update control state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"

  # ─── EVIDENCE ───────────────────────────────────────────────────────────────
...
  /aocs:
    get:
      operationId: listAocs
      tags: [aocs]
      summary: List AOCs & audit reports
      parameters:
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: frameworkId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: AOCs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aoc"
    post:
      operationId: createAoc
      tags: [aocs]
      summary: Upload an AOC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AocInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aoc"
  /aocs/{id}:
    get:
      operationId: getAoc
      tags: [aocs]
      summary: Get an AOC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AOC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aoc"
        "404":
          description: Not found
    patch:
      operationId: updateAoc
      tags: [aocs]
      summary: Update an AOC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AocUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aoc"

  # ─── ASSESSMENTS ────────────────────────────────────────────────────────────
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      parameters:
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: frameworkId
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Not found
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"

  # ─── VENDORS ────────────────────────────────────────────────────────────────
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats for an entity
      parameters:
        - name: entityCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/evidence-by-status:
    get:
      operationId: getEvidenceByStatus
      tags: [dashboard]
      summary: Evidence requests grouped by status
      parameters:
        - name: entityCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence by status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"
  /dashboard/control-coverage:
    get:
      operationId: getControlCoverage
      tags: [dashboard]
      summary: Control coverage by PCI DSS requirement
      parameters:
        - name: entityCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Control coverage breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequirementCoverage"
  /dashboard/upcoming-milestones:
    get:
      operationId: getUpcomingMilestones
      tags: [dashboard]
      summary: Upcoming compliance milestones
      parameters:
        - name: entityCode
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upcoming milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
components:
  schemas:
...
    Control:
      type: object
      required: [id, frameworkId, frameworkCode, ref, title, domain, domainNumber, createdAt]
      properties:
        id:
          type: string
        frameworkId:
          type: string
        frameworkCode:
          type: string
        ref:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        domain:
          type: string
        domainNumber:
          type: number
        entityCode:
          type: ["string", "null"]
        finding:
          type: ["string", "null"]
          enum: [in-place, not-applicable, not-tested, not-in-place, null]
        notes:
          type: ["string", "null"]
        lastTestedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ControlUpdate:
      type: object
      properties:
        finding:
          type: string
          enum: [in-place, not-applicable, not-tested, not-in-place]
        notes:
          type: string
        entityCode:
          type: string

    # ─── Evidence ───────────────────────────────────────────────────────────
...
    Aoc:
      type: object
      required: [id, entityCode, frameworkCode, title, status, createdAt]
      properties:
        id:
          type: string
        entityCode:
          type: string
        frameworkCode:
          type: string
        frameworkName:
          type: ["string", "null"]
        title:
          type: string
        status:
          type: string
          enum: [current, superseded, draft]
        qsaCompany:
          type: ["string", "null"]
        qsaLead:
          type: ["string", "null"]
        issuedDate:
          type: ["string", "null"]
        expiresDate:
          type: ["string", "null"]
        period:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
        controlsCovered:
          type: ["number", "null"]
        findings:
          type: ["number", "null"]
        filename:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AocInput:
      type: object
      required: [entityCode, frameworkCode, title, status]
      properties:
        entityCode:
          type: string
        frameworkCode:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [current, superseded, draft]
        qsaCompany:
          type: string
        qsaLead:
          type: string
        issuedDate:
          type: string
        expiresDate:
          type: string
        period:
          type: string
        result:
          type: string
        controlsCovered:
          type: number
        findings:
          type: number
        filename:
          type: string
        filePath:
          type: string
        notes:
          type: string
    AocUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [current, superseded, draft]
        qsaCompany:
          type: string
        qsaLead:
          type: string
        issuedDate:
          type: string
        expiresDate:
          type: string
        period:
          type: string
        result:
          type: string
        controlsCovered:
          type: number
        findings:
          type: number
        filename:
          type: string
        filePath:
          type: string
        notes:
          type: string

    # ─── Assessment ─────────────────────────────────────────────────────────
    Assessment:
      type: object
      required: [id, entityCode, frameworkId, frameworkCode, name, status, createdAt]
      properties:
        id:
          type: string
        entityCode:
          type: string
        frameworkId:
          type: string
        frameworkCode:
          type: string
        frameworkName:
          type: ["string", "null"]
        name:
          type: string
        qsaCompany:
          type: ["string", "null"]
        leadAssessor:
          type: ["string", "null"]
        plannedStart:
          type: ["string", "null"]
        plannedEnd:
          type: ["string", "null"]
        actualStart:
          type: ["string", "null"]
        actualEnd:
          type: ["string", "null"]
        status:
          type: string
          enum: [planning, fieldwork, reporting, closed]
        createdAt:
          type: string
    AssessmentInput:
      type: object
      required: [entityCode, frameworkId, name]
      properties:
        entityCode:
          type: string
        frameworkId:
          type: string
        name:
          type: string
        qsaCompany:
          type: string
        leadAssessor:
          type: string
        plannedStart:
          type: string
        plannedEnd:
          type: string
        status:
          type: string
          enum: [planning, fieldwork, reporting, closed]
    AssessmentUpdate:
      type: object
      properties:
        name:
          type: string
        qsaCompany:
          type: string
        leadAssessor:
          type: string
        plannedStart:
          type: string
        plannedEnd:
          type: string
        actualStart:
          type: string
        actualEnd:
          type: string
        status:
          type: string
          enum: [planning, fieldwork, reporting, closed]

    # ─── Vendor ─────────────────────────────────────────────────────────────
...
    DashboardSummary:
      type: object
      required: [entityCode, overallReadinessPct, openEvidenceCount, controlsPassing, controlsTotal, overdueEvidenceCount, dueSoonEvidenceCount, nextAocDate, activeFrameworks, activePolicies, openAssessments, vendors]
      properties:
        entityCode:
          type: string
        overallReadinessPct:
          type: number
        openEvidenceCount:
          type: number
        overdueEvidenceCount:
          type: number
        dueSoonEvidenceCount:
          type: number
        controlsPassing:
          type: number
        controlsTotal:
          type: number
        nextAocDate:
          type: ["string", "null"]
        activeFrameworks:
          type: number
        activePolicies:
          type: number
        openAssessments:
          type: number
        vendors:
          type: number
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: number
    RequirementCoverage:
      type: object
      required: [requirementId, requirementName, domain, total, passing, inProgress, blocked, pct]
      properties:
        requirementId:
          type: string
        requirementName:
          type: string
        domain:
          type: string
        total:
          type: number
        passing:
          type: number
        inProgress:
          type: number
        blocked:
          type: number
        pct:
          type: number
    Milestone:
      type: object
      required: [id, title, description, dueDate, priority, category]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
        daysRemaining:
          type: number