# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Compliance OS — GRC tool API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputStatus } from './assessmentInputStatus';

export interface AssessmentInput {
  entityCode: string;
  frameworkId: string;
  name: string;
  qsaCompany?: string;
  leadAssessor?: string;
  plannedStart?: string;
  plannedEnd?: string;
  status?: AssessmentInputStatus;
}