# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Compliance OS — GRC tool API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  entityCode: string;
  overallReadinessPct: number;
  openEvidenceCount: number;
  overdueEvidenceCount: number;
  dueSoonEvidenceCount: number;
  controlsPassing: number;
  controlsTotal: number;
  /** @nullable */
  nextAocDate: string | null;
  activeFrameworks: number;
  activePolicies: number;
  openAssessments: number;
  vendors: number;
}